pip-requirements.txt
django>=1.11,<2.0
djangorestframework
mysqlclient
